requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
scalar-fastapi>=1.0.0
pytest-cov
python-json-logger
httpx
python-multipart
langchain
langchain-community
faiss-cpu
pypdf
langchain-text-splitters
langchain-huggingface
fastembed